# docs/API/openapi/api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Spark API v1
  version: "1.0.0"
  description: >
    Spark's client-neutral, capability-scoped REST API. Hand-written from
    routes/api.php and the underlying controllers on branch
    feature/api-mobile-mcp-parity — not generated, so keep it updated
    alongside docs/API/API_v1.md when routes change. See
    docs/API/README.md for the cross-surface capability model.
...
servers:
  - url: /api/v1
security:
  - sparkToken: []
...
paths:
  /events:
    get:
      operationId: listEvents
      tags: [data]
      x-spark-ability: data:read
      summary: Cursor-paginated reverse-chronological event feed
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100 } }
        - { name: domain, in: query, schema: { type: string } }
        - { name: date, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CompactEvent' } }
                  next_cursor: { type: [string, "null"] }
                  has_more: { type: boolean }
  /events/{id}:
    get:
      operationId: getEvent
      tags: [data]
      x-spark-ability: data:read
      summary: Single event with full embedded blocks
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CompactEvent' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /objects/{id}:
    get:
      operationId: getObject
      tags: [data]
      x-spark-ability: data:read
      summary: Single EventObject, optionally with recent events
      parameters:
        - { $ref: '#/components/parameters/EntityId' }
        - { name: include_events, in: query, schema: { type: boolean, default: true } }
        - { name: event_limit, in: query, schema: { type: integer, minimum: 1, maximum: 25, default: 5 } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CompactObject' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /blocks/{id}:
    get:
      operationId: getBlock
      tags: [data]
      x-spark-ability: data:read
      summary: Single block by UUID
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CompactBlock' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /search:
    get:
      operationId: search
      tags: [data]
      x-spark-ability: data:read
      summary: Multi-mode search across events, objects, integrations, metrics
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: mode, in: query, schema: { type: string, enum: [default, semantic, tag, metric, integration], default: default } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode: { type: string }
                  query: { type: string }
                  events: { type: array, items: { $ref: '#/components/schemas/CompactEvent' } }
                  objects: { type: array, items: { $ref: '#/components/schemas/CompactObject' } }
                  integrations: { type: array, items: { $ref: '#/components/schemas/CompactIntegration' } }
                  metrics: { type: array, items: { $ref: '#/components/schemas/CompactMetric' } }
        "422": { $ref: '#/components/responses/ValidationError' }
  /tags:
    get:
      operationId: listTags
      tags: [data]
      x-spark-ability: data:read
      summary: Cursor-paginated list of the user's tags
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 255 } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 30 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Tag' } }
                  next_cursor: { type: [string, "null"] }
                  has_more: { type: boolean }
  /tags/suggest:
    get:
      operationId: suggestTags
      tags: [data]
      x-spark-ability: data:read
      summary: Autocomplete tag suggestions
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 25, default: 10 } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Tag' } } } } } } }
  /tags/{id}:
    get:
      operationId: getTag
      tags: [data]
      x-spark-ability: data:read
      summary: A single tag plus the items tagged with it
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 30 } }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /map/data:
    get:
      operationId: mapData
      tags: [data]
      x-spark-ability: data:read
      summary: Geo-located events and places within a bounding box (clusters above 500 results)
      parameters:
        - { name: bbox, in: query, required: true, schema: { type: string }, description: "swLat,swLng,neLat,neLng" }
      responses:
        "200": { description: OK }
        "422": { $ref: '#/components/responses/ValidationError' }
  /places/{id}:
    get:
      operationId: getPlace
      tags: [data]
      x-spark-ability: data:read
      summary: A single place (EventObject with concept=place)
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CompactPlace' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /{kind}/{id}/relationships:
    get:
      operationId: listRelationships
      tags: [data]
      x-spark-ability: data:read
      summary: List relationships attached to an owned entity
      parameters:
        - { $ref: '#/components/parameters/EntityKind' }
        - { $ref: '#/components/parameters/EntityId' }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Relationship' } } } } } } }
        "404": { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createRelationship
      tags: [data]
      x-spark-ability: data:write
      summary: Create a relationship from an owned entity to another
      parameters:
        - { $ref: '#/components/parameters/EntityKind' }
        - { $ref: '#/components/parameters/EntityId' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [to_kind, to_id, type]
              properties:
                to_kind: { type: string }
                to_id: { type: string, format: uuid }
                type: { type: string }
                value: { type: [number, "null"] }
                value_multiplier: { type: [number, "null"] }
                value_unit: { type: [string, "null"] }
                metadata: { type: object }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Relationship' } } } }
        "422": { $ref: '#/components/responses/ValidationError' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /relationships/{relationship}:
    delete:
      operationId: deleteRelationship
      tags: [data]
      x-spark-ability: data:write
      summary: Delete an owned relationship
      parameters:
        - { name: relationship, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        "204": { description: No content }
        "404": { $ref: '#/components/responses/NotFound' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
...
  /events/{id}/note:
    patch:
      operationId: updateEventNote
      tags: [data]
      x-spark-ability: data:write
      summary: Set or clear an event's user-authored note
      parameters: [{ $ref: '#/components/parameters/EntityId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { note: { type: [string, "null"] } } }
      responses:
        "200": { description: OK }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /events/{id}/tags:
    post:
      operationId: attachEventTag
      tags: [data]
      x-spark-ability: data:write
      summary: Attach a tag to an owned event
      parameters: [{ $ref: '#/components/parameters/EntityId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tag_id: { type: integer }
                name: { type: string }
                type: { type: string }
      responses:
        "201": { description: Created }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/ValidationError' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /events/{id}/tags/{tagId}:
    delete:
      operationId: detachEventTag
      tags: [data]
      x-spark-ability: data:write
      summary: Detach a tag from an owned event
      parameters:
        - { $ref: '#/components/parameters/EntityId' }
        - { name: tagId, in: path, required: true, schema: { type: integer } }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /objects/{id}/tags:
    post:
      operationId: attachObjectTag
      tags: [data]
      x-spark-ability: data:write
      summary: Attach a tag to an owned object
      parameters: [{ $ref: '#/components/parameters/EntityId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { tag_id: { type: integer }, name: { type: string }, type: { type: string } } }
      responses:
        "201": { description: Created }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/ValidationError' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /objects/{id}/tags/{tagId}:
    delete:
      operationId: detachObjectTag
      tags: [data]
      x-spark-ability: data:write
      summary: Detach a tag from an owned object
      parameters:
        - { $ref: '#/components/parameters/EntityId' }
        - { name: tagId, in: path, required: true, schema: { type: integer } }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /bookmarks:
    post:
      operationId: createBookmark
      tags: [data]
      x-spark-ability: data:write
      summary: Bookmark a URL
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [url], properties: { url: { type: string, format: uri, maxLength: 2048 } } }
      responses:
        "201": { description: Created }
        "200": { description: Already bookmarked }
        "422": { $ref: '#/components/responses/ValidationError' }
  /knowledge/events/{id}/reprocess:
    post:
      operationId: reprocessKnowledgeEvent
      tags: [data]
      x-spark-ability: data:write
      summary: Queue AI reprocessing for a Fetch or Newsletter knowledge event
      parameters: [{ $ref: '#/components/parameters/EntityId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { mode: { type: string, enum: [auto, summary_only, refetch], default: auto } } }
      responses:
        "202": { description: Queued }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/ValidationError' }
        "428": { $ref: '#/components/responses/PreconditionRequired' }
        "412": { $ref: '#/components/responses/PreconditionFailed' }
  /day-summary:
    get:
      operationId: daySummary
      tags: [insights]
      x-spark-ability: insights:read
      summary: Compact, pre-aggregated daily summary with baseline comparison and anomaly detection
      parameters:
        - { name: date, in: query, schema: { type: string, default: today } }
        - { name: domains, in: query, schema: { type: string }, description: "Comma-separated domain filter" }
      responses:
        "200": { description: OK }
...
  /check-ins:
    get:
      operationId: checkInStatus
      tags: [insights]
      x-spark-ability: insights:read
      summary: Morning/afternoon check-in status for a date
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        "200": { description: OK }
    post:
      operationId: submitCheckIn
      tags: [insights]
      x-spark-ability: insights:write
      summary: Submit a check-in (idempotent by period + date)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [period, physical, mental, date]
              properties:
                period: { type: string, enum: [morning, afternoon] }
                physical: { type: integer, minimum: 1, maximum: 5 }
                mental: { type: integer, minimum: 1, maximum: 5 }
                date: { type: string, format: date }
                latitude: { type: number }
                longitude: { type: number }
                address: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CompactEvent' } } } }
  /check-ins/history:
    get:
      operationId: checkInHistory
      tags: [insights]
      x-spark-ability: insights:read
      summary: Check-in history for a date range (max 90 days)
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
      responses:
        "200": { description: OK }
        "422": { $ref: '#/components/responses/ValidationError' }
...
  /up-to-speed/read:
    post:
      operationId: markUpToSpeedRead
      tags: [insights]
      x-spark-ability: insights:write
      summary: Mark Up to Speed items as caught up (idempotent)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: object
                    properties:
                      type: { type: string, enum: [flint_digest, anomaly, news_summary] }
                      id: { type: string, format: uuid }
      responses:
        "200": { description: OK }
...
  /anomalies/{id}/acknowledge:
    post:
      operationId: acknowledgeAnomaly
      tags: [insights]
      x-spark-ability: insights:write
      summary: Acknowledge a metric anomaly
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { note: { type: string, maxLength: 500 }, suppress_until: { type: string, format: date } } }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /integrations/{id}:
    get:
      operationId: getIntegration
      tags: [integrations]
      x-spark-ability: integrations:read
      summary: A single integration
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CompactIntegration' } } } }
        "404": { $ref: '#/components/responses/NotFound' }
  /integrations/{id}/sync:
    post:
      operationId: syncIntegration
      tags: [integrations]
      x-spark-ability: integrations:sync
      summary: Trigger an immediate fetch for one integration instance
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        "200": { description: OK }
        "422": { description: Integration is paused }
  /integrations/sync:
    post:
      operationId: syncIntegrationsByService
      tags: [integrations]
      x-spark-ability: integrations:sync
      summary: Trigger an immediate fetch for every non-paused instance of a service
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [service], properties: { service: { type: string, maxLength: 100 } } }
      responses:
        "200": { description: OK }
        "404": { description: No integrations found for that service }
...
components:
  securitySchemes:
    sparkToken:
      type: http
      scheme: bearer
      description: >
        Laravel Sanctum personal access token. Each operation's x-spark-ability
        extension names the ability the token must carry
        (App\Support\SparkAbility::allows) — OpenAPI has no native primitive
        for Sanctum's per-operation string abilities.
...
  schemas:
...
    CompactEvent:
      type: object
      description: App\Http\Resources\Compact\CompactEventResource
      properties:
        id: { type: string, format: uuid }
        time: { type: string, format: date-time }
        service: { type: string }
        domain: { type: string }
        action: { type: string }
        value: { type: string }
        unit: { type: string }
        display_value: { type: string }
        url: { type: string }
        actor: { type: object }
        target: { type: object }
        tags: { type: array, items: { type: object } }
        tldr: { type: string }
        note: { type: string }
        blocks_count: { type: integer }
        blocks: { type: array, items: { $ref: '#/components/schemas/CompactBlock' } }
        display_name: { type: string }
        display_with_object: { type: boolean }
        hidden: { type: boolean }
...